+++ app/issues/new/page.tsx
@@ -1,83 +1,10 @@
-'use client';
-
-import ErrorMessage from '@/app/components/ErrorMessage';
-import Spinner from '@/app/components/Spinner';
-import { createIssueSchema } from '@/app/validationSchema';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Button, Callout, TextField } from '@radix-ui/themes';
-import axios from 'axios';
-import 'easymde/dist/easymde.min.css';
-import dynamic from 'next/dynamic';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { z } from 'zod';
-
-// prettier-ignore
-const SimpleMDE = dynamic(
-  () => import('react-simplemde-editor'),
-  { ssr: false }
-);
-
-type IssueForm = z.infer<typeof createIssueSchema>;
+import React from 'react'
+import IssueForm from '../_components/IssueForm'
 
 const NewIssuePage = () => {
-  const router = useRouter();
-  const [error, setError] = useState('');
-  const [isSubmitting, setSubmitting] = useState(false);
-
-  const {
-    register,
-    control,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<IssueForm>({
-    resolver: zodResolver(createIssueSchema),
-  });
-
-  const onSubmit = handleSubmit(async (data: IssueForm) => {
-    try {
-      setSubmitting(true);
-
-      await axios.post('/api/issues', data);
-      router.push('/issues');
-    } catch (error) {
-      setSubmitting(false);
-
-      setError('An unexpected error occurred.');
-    }
-  });
-
   return (
-    <div className="max-w-xl">
-      {error && (
-        <Callout.Root color="red" className="mb-3" size="1">
-          <Callout.Text>{error}</Callout.Text>
-        </Callout.Root>
-      )}
-      <form className=" space-y-3" onSubmit={onSubmit}>
-        <TextField.Root>
-          <TextField.Input placeholder="Title " {...register('title')} />
-        </TextField.Root>
+    <IssueForm />
+  )
+}
 
-        <ErrorMessage>{errors.title?.message}</ErrorMessage>
-
-        <Controller
-          name="description"
-          control={control}
-          render={({ field }) => (
-            <SimpleMDE placeholder="Description" {...field} />
-          )}
-        />
-
-        <ErrorMessage>{errors.description?.message}</ErrorMessage>
-
-        <Button disabled={isSubmitting}>
-          Submit New Issue {isSubmitting && <Spinner />}
-        </Button>
-      </form>
-    </div>
-  );
-};
-
-export default NewIssuePage;
+export default NewIssuePage
